discord_event_test.py
import os
import discord

DISCORD_BOT_TOKEN = os.getenv("DISCORD_BOT_TOKEN")
SERVER_ID = os.getenv("SERVER_ID")
CHANNEL_ID = int(os.getenv("CHANNEL_ID"))



intents = discord.Intents.default()
intents.members = True

client = discord.Client(intents=intents)
@client.event
async def on_ready():
    print(f"Logged in as {client.user}")
    channel = client.get_channel(CHANNEL_ID)  # Fetch the channel object
    await channel.send("/overview server")
    await client.close()

# Log in to Discord runs the on_ready function
client.run(DISCORD_BOT_TOKEN)