requirements.txt
pytz
boto3
requests
icalendar
beautifulsoup4
discord.py
aiohttp
